Always checkout the refspec given
In some edge cases the refspec given will not be valid
for the manifest we're using (i.e. uses the wrong branch).In this
case we should just use it anyway.

Also checks that at least one change has been found and adds further
error checking to fetch_project.sh and fetch_project.bat in case of
an empty line.

Reviewed-on: http://review.couchbase.org/60494
Tested-by: Will Gardner <will.gardner@couchbase.com>

+++ scripts/jenkins/commit_validation/fetch_project.sh
@@ -18,6 +18,18 @@ if [ -z "$GERRIT_PORT" ]; then
     echo "Error: Required environment variable 'GERRIT_PORT' not set."
     exit 2
 fi
+if [ -z "$PROJECT" ]; then
+    echo "Error: Required environment variable 'PROJECT' not set."
+    exit 3
+fi
+if [ -z "$PROJECT_PATH" ]; then
+    echo "Error: Required environment variable 'PROJECT_PATH' not set."
+    exit 4
+fi
+if [ -z "$REFSPEC" ]; then
+    echo "Error: Required environment variable 'REFSPEC' not set."
+    exit 5
+fi
 
 cd $PROJECT_PATH
 git reset --hard HEAD
